chore(go.plugin): rename runtimemgr for clarity
