+++ src/go/plugin/agent/runtimechartemit/components.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package runtimemgr
+package runtimechartemit
 
 import (
 	"fmt"
